raise version, remove duplicated code
