+++ client/package.json
@@ -1,9 +1,9 @@
 {
   "name": "eventbox-dashboard",
-  "version": "0.1.0",
+  "version": "0.9.0",
   "private": true,
   "dependencies": {
-    "@babel/plugin-proposal-decorators": "^7.3.0",
+    "@babel/plugin-proposal-decorators": "^7.4.0",
     "@types/jest": "^24.0.9",
     "@types/node": "^11.11.3",
     "@types/react": "^16.8.6",
@@ -11,7 +11,7 @@
     "@types/react-router-dom": "^4.3.1",
     "ag-grid-community": "^20.1.0",
     "ag-grid-react": "^20.1.1",
-    "antd": "^3.15.0",
+    "antd": "^3.15.1",
     "apollo-cache-inmemory": "^1.5.1",
     "apollo-client": "^2.5.1",
     "apollo-link": "^1.2.11",
@@ -49,8 +49,8 @@
     "react-app-rewired": "^2.1.1",
     "react-dom": "^16.8.4",
     "react-draft-wysiwyg": "^1.13.2",
-    "react-i18next": "^10.5.1",
-    "react-router-dom": "^4.4.0",
+    "react-i18next": "^10.5.2",
+    "react-router-dom": "^5.0.0",
     "react-scripts": "^2.1.8",
     "react-scroll": "^1.7.11",
     "subscriptions-transport-ws": "^0.9.16",
